made an Interface for IslandProperties & removed an no longer available method
